Fix bug: Single-instance buttons do nothing after disabled
This is to fix open-craft/problem-builder#118, where a button marked with
'single-instance' data got disabled, hence its click event handler at
edx-platform is not called.

This patch solves this issue by correcting the calling order, so that we
invoke the click event handler at edx-platform first; then we disable
the button.