build: Update upgrade-python-requirements.yml
The $default_branch value was supposed to get rendered when the workflow was copied from the workflow-template but it didn't get updated. Update the default value to a string.

We're also trying an experiment where we grab the default branch
dynamically for the runs not called via workflow-dispatch.

temp: testing a different way.

temp: testing a different way.